uprgades

+++ ft_pf_putnbr_hex.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_pf_putnbr_hex.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hfilipe- <hfilipe-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: hfilipe- <hfilipe-@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 09:20:03 by hfilipe-          #+#    #+#             */
-/*   Updated: 2024/11/22 15:09:51 by hfilipe-         ###   ########.fr       */
+/*   Updated: 2024/11/22 19:07:43 by hfilipe-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "printf.h"
 
-size_t	ft_pf_putnbr_hex(int n, char *base)
+size_t	ft_pf_putnbr_hex(unsigned long n, char *base)
 {
 	size_t	numb_char;
 	int		array[20];
