+++ ft_pf_putnbr_hex_fw.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_pf_putnbr_hex_fw.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hfilipe- <hfilipe-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: hfilipe- <hfilipe-@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/22 15:03:50 by hfilipe-          #+#    #+#             */
-/*   Updated: 2024/11/22 15:33:04 by hfilipe-         ###   ########.fr       */
+/*   Updated: 2024/11/22 19:07:35 by hfilipe-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "printf.h"
 
-size_t	ft_pf_putnbr_hex_fw(int n, char *base, size_t field_len)
+size_t	ft_pf_putnbr_hex_fw(unsigned long n, char *base, size_t field_len)
 {
 	size_t	numb_char;
 	int		array[11];
